reregister belly button block and item

+++ src/main/java/teamdraco/bellybutton/common/blocks/BellyButtonBlock.java
@@ -32,11 +32,11 @@ public class BellyButtonBlock extends ButtonBlock {
             player.playNotifySound(SoundEvents.WOOL_BREAK, SoundSource.BLOCKS, 0.8f, 1.0f);
             player.swing(handIn, true);
             if (this.RANDOM.nextInt(50) == 0) {
-                ItemEntity itemEntity = new ItemEntity(player.getCommandSenderWorld(), (double) pos.getX() + 0.5D, (double) (pos.getY() + 0.5D), (double) pos.getZ() + 0.5D, new ItemStack(BellyButtonItems.LINT.get(), 1));
+                ItemEntity itemEntity = new ItemEntity(player.getCommandSenderWorld(), (double) pos.getX() + 0.5D, pos.getY() + 0.5D, (double) pos.getZ() + 0.5D, new ItemStack(BellyButtonItems.LINT.get(), 1));
                 player.getCommandSenderWorld().addFreshEntity(itemEntity);
             }
             if (this.RANDOM.nextInt(5000) == 0) {
-                ItemEntity itemEntity = new ItemEntity(player.getCommandSenderWorld(), (double) pos.getX() + 0.5D, (double) (pos.getY() + 0.5D), (double) pos.getZ() + 0.5D, new ItemStack(BellyButtonItems.MUSIC_DISC_BELLY_BOPPIN.get(), 1));
+                ItemEntity itemEntity = new ItemEntity(player.getCommandSenderWorld(), (double) pos.getX() + 0.5D, pos.getY() + 0.5D, (double) pos.getZ() + 0.5D, new ItemStack(BellyButtonItems.MUSIC_DISC_BELLY_BOPPIN.get(), 1));
                 player.getCommandSenderWorld().addFreshEntity(itemEntity);
             }
         }
